# src/backend/openapi.yaml
openapi: 3.1.0
info:
  title: LiDAR Forest Analysis API
  version: 1.0.0
  description: |
    REST API for the LiDAR Forest Analysis Platform. Transform LiDAR point cloud data
    into professional forest inventory reports, carbon stock estimates, and 3D visualizations.

    ## Authentication

    All API requests require authentication via API key. Include your API key in requests using
    one of these methods:

    - **Authorization header** (recommended): `Authorization: Bearer lf_live_xxx`
    - **X-API-Key header**: `X-API-Key: lf_live_xxx`

    ## Rate Limits

    Rate limits vary by tier:

    | Tier | Requests/Minute | Requests/Day |
    |------|-----------------|--------------|
    | Free | 60 | 10,000 |
    | Starter | 120 | 50,000 |
    | Professional | 300 | 200,000 |
    | Enterprise | 1,000 | Unlimited |

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Max requests per minute
    - `X-RateLimit-Remaining`: Remaining requests this minute
    - `X-RateLimit-Reset`: Time when the rate limit resets (ISO 8601)
  contact:
    name: API Support
    email: api@lidarforest.com
    url: https://docs.lidarforest.com
  license:
    name: Proprietary
    url: https://lidarforest.com/terms
servers:
  - url: https://api.lidarforest.com/api/v1
    description: Production server
  - url: https://staging-api.lidarforest.com/api/v1
    description: Staging server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /projects:
    get:
      summary: List projects
      tags: [Projects]
      operationId: listProjects
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, name]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
    post:
      summary: Create project
      tags: [Projects]
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectInput'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /projects/{projectId}:
    get:
      summary: Get project
      tags: [Projects]
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update project
      tags: [Projects]
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectInput'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete project
      tags: [Projects]
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # Files
  /files:
    get:
      summary: List files
      tags: [Files]
      operationId: listFiles
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
  /files/upload-url:
    post:
      summary: Get upload URL
      tags: [Files]
      operationId: getUploadUrl
      description: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlInput'
      responses:
        '201':
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
  /files/{fileId}:
    get:
      summary: Get file details
      tags: [Files]
      operationId: getFile
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: File details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete file
      tags: [Files]
      operationId: deleteFile
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: File deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /files/{fileId}/download-url:
    get:
      summary: Get download URL
      tags: [Files]
      operationId: getDownloadUrl
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: Download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlResponse'

  # Analyses
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key as Bearer token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key header
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    AnalysisType:
      type: string
      enum: [TREE_DETECTION, SPECIES_CLASSIFICATION, CARBON_ESTIMATE, FULL_INVENTORY]
...
    Status:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED]

    # Request schemas
    CreateProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        location:
          type: string
          maxLength: 500
        metadata:
          type: object
    UpdateProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
        location:
          type: string
        metadata:
          type: object
    UploadUrlInput:
      type: object
      required: [projectId, filename, fileSize]
      properties:
        projectId:
          type: string
          format: uuid
        filename:
          type: string
          minLength: 1
          maxLength: 255
        fileSize:
          type: integer
          minimum: 1
        mimeType:
          type: string
        metadata:
          type: object
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        location:
          type: string
        status:
          type: string
        fileCount:
          type: integer
        analysisCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ProjectResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Project'
    ProjectDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Project'
            - type: object
              properties:
                files:
                  type: array
                  items:
                    $ref: '#/components/schemas/FileSummary'
                analyses:
                  type: array
                  items:
                    $ref: '#/components/schemas/AnalysisSummary'
    FileSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        fileSize:
          type: integer
        status:
          $ref: '#/components/schemas/Status'
        createdAt:
          type: string
          format: date-time
    File:
      allOf:
        - $ref: '#/components/schemas/FileSummary'
        - type: object
          properties:
            projectId:
              type: string
              format: uuid
            originalFilename:
              type: string
            mimeType:
              type: string
            metadata:
              type: object
    FileListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/File'
        pagination:
          $ref: '#/components/schemas/Pagination'
    FileResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/File'
    UploadUrlResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            fileId:
              type: string
              format: uuid
            uploadUrl:
              type: string
              format: uri
            expiresAt:
              type: string
              format: date-time
            instructions:
              type: object
              properties:
                method:
                  type: string
                headers:
                  type: object
    DownloadUrlResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            fileId:
              type: string
              format: uuid
            filename:
              type: string
            downloadUrl:
              type: string
              format: uri
            expiresAt:
              type: string
              format: date-time
    AnalysisSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/AnalysisType'
        status:
          $ref: '#/components/schemas/Status'
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  hint:
                    type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  retryAfter:
                    type: integer
                  limits:
                    type: object